+++ src/main/open-window.ts
@@ -179,7 +179,7 @@ export class OpenWindow {
             );
         }
 
-        // this.window.webContents.openDevTools();
+        this.window.webContents.openDevTools();
     }
 
     send(eventName: EventEndpoint, data: any) {